feat: change hoursOfHeating to ints from decimals as this is the resolution of the request bre accepts

+++ SeaPublicWebsite.UnitTests/QuestionFlowServiceTests.cs
@@ -1742,8 +1742,8 @@ public class QuestionFlowServiceTests
             HasHotWaterCylinder? hasHotWaterCylinder = null,
             int? numberOfOccupants = null,
             HeatingPattern? heatingPattern = null,
-            decimal? hoursOfHeatingMorning = null,
-            decimal? hoursOfHeatingEvening = null,
+            int? hoursOfHeatingMorning = null,
+            int? hoursOfHeatingEvening = null,
             decimal? temperature = null,
             QuestionFlowPage? entryPoint = null)
         {
